docker/tenkobot/pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["zankoo <zhuqi1101@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-alpha.15"
nonebot-adapter-cqhttp = "^2.0.0-alpha.15"
bilibili-api = "^8.2.0"
nonebot-plugin-apscheduler = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[nonebot.plugins]
plugins = []
plugin_dirs = ["src/plugins"]